pyproject.toml
[project]
name = "realest-tate-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Ivan Pupkin", email = "ipupkin@untref.edu.ar" }
]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.12",
    "ipython>=9.3.0",
    "langchain>=0.3.25",
    "langchain-aws>=0.2.24",
    "langgraph-supervisor>=0.0.27",
    "mlflow>=2.22.1",
    "psycopg2-binary>=2.9.10",
    "pyppeteer>=2.0.0",
    "python-dotenv>=1.1.0",
]

[project.scripts]
realest-tate-agent = "realest_tate_agent:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
